DTR-376: Adding test cases

+++ app/uk/gov/hmrc/rdsdatacacheproxy/models/requests/PaymentPlanDuplicateCheckRequest.scala
@@ -24,8 +24,8 @@ case class PaymentPlanDuplicateCheckRequest(
                                              planType: String,
                                              paymentService: String,
                                              paymentReference: String,
-                                             paymentAmount: BigDecimal,
-                                             totalLiability: BigDecimal,
+                                             paymentAmount: Double,
+                                             totalLiability: Double,
                                              paymentFrequency: String
                                            )
